pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-automate"
version = "0.1.0"
description = "Automate the Azure DevOps process from intake to epic, feature, user story and tasks with AI, GitHub and quality standards"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "azure-devops>=7.1.0b4",
    "msrest>=0.7.1",
    "openai>=1.30.0",
    "PyGithub>=2.3.0",
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "pydantic>=2.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
devops-automate = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
